+++ src/api/permissions.h
@@ -15,10 +15,10 @@
 extern "C" {
 #endif
 
-extern uint8_t base_run(void *args);
-extern uint8_t base_release(uint8_t error);
-extern uint8_t get_invocation_args(uint32_t *pid, void **args);
-extern uint8_t grant_permission(uint32_t pid, uint16_t permission);
+extern uint32_t base_run(void *args);
+extern uint32_t base_release(uint8_t error);
+extern uint32_t get_invocation_args(uint32_t *pid, void **args);
+extern uint32_t grant_permission(uint32_t pid, uint16_t permission);
 
 #ifdef __cplusplus
 }
